fix: re-organize utils packages and don't use ioutils anymore
