+++ internal/utils/results/parse.go
@@ -1,11 +1,52 @@
-package parse
+package results
 
 import (
 	"hypha/api/internal/db"
-	"hypha/api/internal/utils/results/structs"
 	"strings"
 	"time"
 )
+
+// ParseJUnitResults parses JUnit test results and stores them in the database.
+//
+// This function iterates over the provided JUnit test suites, creates corresponding
+// result and test suite models, and saves them to the database. It also creates and
+// saves properties and test cases associated with each test suite.
+//
+// Parameters:
+// - testSuites: The JUnitTestSuites containing the test results to be parsed.
+// - dbOperations: The DatabaseOperations interface for interacting with the database.
+// - productId: The ID of the product for which the test results are being parsed.
+//
+// Returns:
+// - error: An error if there is any issue during the parsing or saving of the test results.
+func ParseJUnitResults(testSuites JUnitTestSuites, dbOperations db.DatabaseOperations, productId string) error {
+	for _, suite := range testSuites.TestSuites {
+		resultModel, err := createResultModel(productId)
+		if err != nil {
+			return err
+		}
+		if err := dbOperations.Create(&resultModel); err != nil {
+			return err
+		}
+
+		testSuiteModel, err := createTestSuiteModel(suite, resultModel.ID)
+		if err != nil {
+			return err
+		}
+		if err := dbOperations.Create(&testSuiteModel); err != nil {
+			return err
+		}
+
+		if err := createAndSaveProperties(suite.Properties, testSuiteModel.ID, dbOperations); err != nil {
+			return err
+		}
+
+		if err := createAndSaveTestCases(suite.TestCases, testSuiteModel.ID, dbOperations); err != nil {
+			return err
+		}
+	}
+	return nil
+}
 
 // trimLeadingWhitespace removes the leading whitespace from each line of the input text.
 // It calculates the minimum indentation level across all lines and removes that amount
@@ -86,7 +127,7 @@ func createResultModel(productId string) (db.Result, error) {
 // Returns:
 // - db.TestSuite: The created TestSuite model.
 // - error: An error if there is any issue during the creation of the model.
-func createTestSuiteModel(suite structs.JUnitTestSuite, resultID string) (db.TestSuite, error) {
+func createTestSuiteModel(suite JUnitTestSuite, resultID string) (db.TestSuite, error) {
 	return db.TestSuite{
 		ID:         db.GenerateUniqueID(),
 		ResultID:   resultID,
@@ -113,7 +154,7 @@ func createTestSuiteModel(suite structs.JUnitTestSuite, resultID string) (db.Tes
 //
 // Returns:
 // - error: An error if there is any issue during the creation or saving of the properties.
-func createAndSaveProperties(properties []structs.Property, testSuiteID string, dbOperations db.DatabaseOperations) error {
+func createAndSaveProperties(properties []Property, testSuiteID string, dbOperations db.DatabaseOperations) error {
 	for _, property := range properties {
 		value := property.Value
 		if value == "" {
@@ -143,7 +184,7 @@ func createAndSaveProperties(properties []structs.Property, testSuiteID string, 
 //
 // Returns:
 // - error: An error if there is any issue during the creation or saving of the test cases or their properties.
-func createAndSaveTestCases(testCases []structs.JUnitTestCase, testSuiteID string, dbOperations db.DatabaseOperations) error {
+func createAndSaveTestCases(testCases []JUnitTestCase, testSuiteID string, dbOperations db.DatabaseOperations) error {
 	for _, testCase := range testCases {
 		testCaseModel, err := createTestCaseModel(testCase, testSuiteID)
 		if err != nil {
@@ -170,7 +211,7 @@ func createAndSaveTestCases(testCases []structs.JUnitTestCase, testSuiteID strin
 // Returns:
 // - db.TestCase: The created TestCase model.
 // - error: An error if there is any issue during the creation of the model.
-func createTestCaseModel(testCase structs.JUnitTestCase, testSuiteID string) (db.TestCase, error) {
+func createTestCaseModel(testCase JUnitTestCase, testSuiteID string) (db.TestCase, error) {
 	status, message, testCaseType := determineTestCaseStatus(testCase)
 
 	return db.TestCase{
@@ -200,7 +241,7 @@ func createTestCaseModel(testCase structs.JUnitTestCase, testSuiteID string) (db
 // - string: The status of the test case ("pass", "fail", "error", or "skipped").
 // - *string: The message associated with the test case status, if any.
 // - *string: The type of the test case status, if any.
-func determineTestCaseStatus(testCase structs.JUnitTestCase) (string, *string, *string) {
+func determineTestCaseStatus(testCase JUnitTestCase) (string, *string, *string) {
 	status := "pass"
 	var message *string
 	var testCaseType *string
@@ -232,7 +273,7 @@ func determineTestCaseStatus(testCase structs.JUnitTestCase) (string, *string, *
 //
 // Returns:
 // - error: An error if there is any issue during the creation or saving of the properties.
-func createAndSaveTestCaseProperties(properties []structs.Property, testCaseID string, dbOperations db.DatabaseOperations) error {
+func createAndSaveTestCaseProperties(properties []Property, testCaseID string, dbOperations db.DatabaseOperations) error {
 	for _, property := range properties {
 		value := property.Value
 		if value == "" {
